+++ src/Timeout.ts
@@ -1,5 +1,7 @@
 import { validateTime } from "./validateTime.js";
 
+type Nil = null | undefined;
+
 /**
  * Replacement class for `setTimeout`
  */
@@ -8,12 +10,14 @@ class Timeout {
 
     private _time: number;
 
+    private _started = 0;
+
     private _timerId: ReturnType<typeof setTimeout> | null;
 
     /**
-     * @param {function} callback - function to be called when given time passes
-     * @param {number} time - time in ms to fire the callback
-     * @param {boolean} [start] - start the timer
+     * @param callback - function to be called when given time passes
+     * @param time - time in ms to fire the callback
+     * @param start - start the timer
      */
     public constructor(callback: () => void, time: number, start: boolean = false) {
         validateTime(time);
@@ -26,15 +30,13 @@ class Timeout {
     }
 
     /**
-     * Starts or restarts the timer
+     * Starts or restarts the timer.
      *
-     * @param {number} [newTime] - override time to call the callback
+     * @param newTime - override time to call the callback
      */
-    public start(newTime?: number) {
-        if (newTime != null) {
-            validateTime(newTime);
-            this._time = newTime;
-        }
+    public start(newTime?: number | Nil) {
+        this._updateTime(newTime);
+
         this.stop();
         this._start();
     }
@@ -45,12 +47,40 @@ class Timeout {
                 this._cb();
                 this.stop();
             }, this._time);
+            this._started = Date.now();
         }
     }
 
+    private _updateTime(newTime: number | Nil) {
+        if (newTime == null) {
+            return;
+        }
+        validateTime(newTime);
+        this._time = newTime;
+    }
+
     /**
-     * Starts the timer only if it's not already started
-     * @returns {boolean} - true if newly started, false if already started
+     * Changes the time to call the callback.
+     * If the timer is started, it will be restarted (but not when the same time is given, in this case, the timer will continue).
+     * If the timer is not started, new time will be used when it's started.
+     *
+     * @param newTime - time to call the callback
+     */
+    public changeTime(newTime: number) {
+        if (this._time === newTime) {
+            return;
+        }
+        this._updateTime(newTime);
+        this.restartOnly(newTime);
+    }
+
+    /**
+     * Starts the timer only if it's not already started.
+     *
+     * This function does not allow you to change the time while calling it to avoid any ambiguity about if the timer
+     * should be restarted as usual when changing the time or not.
+     *
+     * @returns true if newly started, false if already started
      */
     public startOnly() {
         if (this._timerId !== null) {
@@ -61,10 +91,15 @@ class Timeout {
     }
 
     /**
-     * Restarts the timer only if it's already started
-     * @returns {boolean} - true if restarted, false if not started
+     * Restarts the timer only if it's already started.
+     * If newTime is given and the timer is stopped, it will be saved and used when the timer is started.
+     *
+     * @param newTime - override time to call the callback
+     * @returns true if restarted, false if not started
      */
-    public restartOnly() {
+    public restartOnly(newTime?: number | Nil) {
+        this._updateTime(newTime);
+
         if (this._timerId === null) {
             return false;
         }
@@ -83,8 +118,21 @@ class Timeout {
         }
     }
 
+    /**
+     * @returns true if the timer is started, false otherwise
+     */
     public get started() {
         return this._timerId !== null;
+    }
+
+    /**
+     * @returns time left in ms
+     */
+    public get timeLeft() {
+        if (this._timerId === null) {
+            return 0;
+        }
+        return this._started + this._time - Date.now();
     }
 }