+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+The format is based on [EZEZ Changelog](https://ezez.dev/changelog/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
-(nothing yet)
+### Added
+- `timeLeft` property
+### Changed
+- `restartOnly` methods supports updating timeout and instant first run
+- `startOnly` methods supports updating instant first run
+### Fixed
+- JSDoc is invalid about return types
 
 ## [5.0.0] - 2023-05-07
 ### Added